backend functional commit

+++ controllers/contact-controller.js
@@ -5,7 +5,7 @@ const Contact = require('../models/contactModel');
 //@route GET /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req,res)=> {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 })
 
@@ -23,6 +23,7 @@ const createContact = asyncHandler(async (req,res)=> {
         name,
         email,
         phone,
+        user_id: req.user.id,
     })
     res.status(201).json(contact)
 
@@ -50,6 +51,11 @@ const updateContact = asyncHandler(async (req,res)=> {
         res.status(400);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("This user doesn't have permission to update other user contacts!!");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -66,6 +72,10 @@ const deleteContact = asyncHandler(async (req,res)=> {
     if(!contact){
         res.status(404);
         throw new Error("Contact not found")
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("This user doesn't have permission to delete other user contacts!!")
     }
     await Contact.findByIdAndDelete(contact)
     res.status(200).json(contact)
